# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MediaTracker",
    "version": "0.1.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "security": [
    {
      "ApiKey": []
    },
    {
      "ApiHeader": []
    }
  ],
  "paths": {
...
    "/api/list": {
      "put": {
        "tags": [
          "List"
        ],
        "operationId": "Add list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "privacy": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ListPrivacy"
                      }
                    ],
                    "nullable": true
                  },
                  "sortBy": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ListSortBy"
                      }
                    ],
                    "nullable": true
                  },
                  "sortOrder": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ListSortOrder"
                      }
                    ],
                    "nullable": true
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/List"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "List"
        ],
        "operationId": "Update list",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "privacy": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ListPrivacy"
                      }
                    ],
                    "nullable": true
                  },
                  "sortBy": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ListSortBy"
                      }
                    ],
                    "nullable": true
                  },
                  "sortOrder": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/ListSortOrder"
                      }
                    ],
                    "nullable": true
                  }
                },
                "required": [
                  "id",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/List"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "List"
        ],
        "operationId": "Get list",
        "parameters": [
          {
            "name": "listId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDetailsResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "List"
        ],
        "operationId": "Delete list",
        "parameters": [
          {
            "name": "listId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListDetailsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/list/items": {
      "get": {
        "tags": [
          "List"
        ],
        "operationId": "Get List Items",
        "parameters": [
          {
            "name": "listId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ListSortBy"
                }
              ],
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListItemsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "ApiHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "MediaType": {
        "enum": [
          "audiobook",
          "book",
          "movie",
          "tv",
          "video_game"
        ],
        "type": "string"
      },
...
      "AudibleCountryCode": {
        "enum": [
          "au",
          "ca",
          "de",
          "es",
          "fr",
          "in",
          "it",
          "jp",
          "uk",
          "us"
        ],
        "type": "string"
      },
...
      "TvSeason": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "numberOfEpisodes": {
            "type": "number",
            "nullable": true
          },
          "externalPosterUrl": {
            "type": "string",
            "nullable": true
          },
          "posterId": {
            "type": "string",
            "nullable": true
          },
          "releaseDate": {
            "type": "string",
            "nullable": true
          },
          "tvShowId": {
            "type": "number",
            "nullable": true
          },
          "tmdbId": {
            "type": "number",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "seasonNumber": {
            "type": "number"
          },
          "tvmazeId": {
            "type": "number",
            "nullable": true
          },
          "episodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TvEpisode"
            },
            "nullable": true
          },
          "userRating": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRating"
              }
            ],
            "nullable": true
          },
          "seen": {
            "type": "boolean",
            "nullable": true
          },
          "posterSmall": {
            "type": "string",
            "nullable": true
          },
          "isSpecialSeason": {
            "type": "boolean"
          },
          "tvdbId": {
            "type": "number",
            "nullable": true
          },
          "traktId": {
            "type": "number",
            "nullable": true
          },
          "onWatchlist": {
            "type": "boolean",
            "nullable": true
          },
          "totalRuntime": {
            "type": "number",
            "nullable": true
          },
          "lastSeenAt": {
            "type": "number",
            "nullable": true
          },
          "unseenEpisodesCount": {
            "type": "number",
            "nullable": true
          },
          "metadataLanguage": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "isSpecialSeason",
          "seasonNumber",
          "title"
        ]
      },
      "TvEpisode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "episodeNumber": {
            "type": "number"
          },
          "seasonNumber": {
            "type": "number"
          },
          "releaseDate": {
            "type": "string",
            "nullable": true
          },
          "tvShowId": {
            "type": "number",
            "nullable": true
          },
          "seasonId": {
            "type": "number",
            "nullable": true
          },
          "tmdbId": {
            "type": "number",
            "nullable": true
          },
          "imdbId": {
            "type": "string",
            "nullable": true
          },
          "runtime": {
            "type": "number",
            "nullable": true
          },
          "seenHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Seen"
            },
            "nullable": true
          },
          "userRating": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRating"
              }
            ],
            "nullable": true
          },
          "lastSeenAt": {
            "type": "number",
            "nullable": true
          },
          "seasonAndEpisodeNumber": {
            "type": "number",
            "nullable": true
          },
          "seen": {
            "type": "boolean",
            "nullable": true
          },
          "tvShow": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaItemItemsResponse"
              }
            ],
            "nullable": true
          },
          "isSpecialEpisode": {
            "type": "boolean"
          },
          "tvdbId": {
            "type": "number",
            "nullable": true
          },
          "traktId": {
            "type": "number",
            "nullable": true
          },
          "onWatchlist": {
            "type": "boolean",
            "nullable": true
          },
          "metadataLanguage": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "episodeNumber",
          "isSpecialEpisode",
          "seasonNumber",
          "title"
        ]
      },
      "Seen": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true
          },
          "date": {
            "type": "number",
            "nullable": true
          },
          "mediaItemId": {
            "type": "number"
          },
          "episodeId": {
            "type": "number",
            "nullable": true
          },
          "userId": {
            "type": "number"
          },
          "duration": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "mediaItemId",
          "userId"
        ]
      },
      "UserRating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true
          },
          "mediaItemId": {
            "type": "number"
          },
          "date": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          },
          "rating": {
            "type": "number",
            "nullable": true
          },
          "review": {
            "type": "string",
            "nullable": true
          },
          "episodeId": {
            "type": "number",
            "nullable": true
          },
          "seasonId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "date",
          "mediaItemId",
          "userId"
        ]
      },
      "MediaItemItemsResponse": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "nullable": true
              },
              "releaseDate": {
                "type": "string",
                "nullable": true
              },
              "posterId": {
                "type": "string",
                "nullable": true
              },
              "backdropId": {
                "type": "string",
                "nullable": true
              },
              "tmdbId": {
                "type": "number",
                "nullable": true
              },
              "imdbId": {
                "type": "string",
                "nullable": true
              },
              "tvdbId": {
                "type": "number",
                "nullable": true
              },
              "tvmazeId": {
                "type": "number",
                "nullable": true
              },
              "igdbId": {
                "type": "number",
                "nullable": true
              },
              "openlibraryId": {
                "type": "string",
                "nullable": true
              },
              "audibleId": {
                "type": "string",
                "nullable": true
              },
              "traktId": {
                "type": "number",
                "nullable": true
              },
              "goodreadsId": {
                "type": "number",
                "nullable": true
              },
              "numberOfSeasons": {
                "type": "number",
                "nullable": true
              },
              "status": {
                "type": "string",
                "nullable": true
              },
              "platform": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "nullable": true
              },
              "title": {
                "type": "string"
              },
              "originalTitle": {
                "type": "string",
                "nullable": true
              },
              "externalPosterUrl": {
                "type": "string",
                "nullable": true
              },
              "externalBackdropUrl": {
                "type": "string",
                "nullable": true
              },
              "tmdbRating": {
                "type": "number",
                "nullable": true
              },
              "overview": {
                "type": "string",
                "nullable": true
              },
              "lastTimeUpdated": {
                "type": "number",
                "nullable": true
              },
              "source": {
                "type": "string"
              },
              "network": {
                "type": "string",
                "nullable": true
              },
              "url": {
                "type": "string",
                "nullable": true
              },
              "runtime": {
                "type": "number",
                "nullable": true
              },
              "mediaType": {
                "$ref": "#/components/schemas/MediaType"
              },
              "genres": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "nullable": true
              },
              "numberOfEpisodes": {
                "type": "number",
                "nullable": true
              },
              "developer": {
                "type": "string",
                "nullable": true
              },
              "director": {
                "type": "string",
                "nullable": true
              },
              "creator": {
                "type": "string",
                "nullable": true
              },
              "publisher": {
                "type": "string",
                "nullable": true
              },
              "authors": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "nullable": true
              },
              "narrators": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "nullable": true
              },
              "language": {
                "type": "string",
                "nullable": true
              },
              "numberOfPages": {
                "type": "number",
                "nullable": true
              },
              "audibleCountryCode": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AudibleCountryCode"
                  }
                ],
                "nullable": true
              },
              "needsDetails": {
                "type": "boolean",
                "nullable": true
              },
              "platformRating": {
                "type": "number",
                "nullable": true
              }
            },
            "required": [
              "mediaType",
              "source",
              "title"
            ]
          },
          {
            "type": "object",
            "properties": {
              "isSearchResult": {
                "type": "boolean",
                "nullable": true
              },
              "hasDetails": {
                "type": "boolean",
                "nullable": true
              },
              "poster": {
                "type": "string",
                "nullable": true
              },
              "posterSmall": {
                "type": "string",
                "nullable": true
              },
              "backdrop": {
                "type": "string",
                "nullable": true
              },
              "seenHistory": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Seen"
                },
                "nullable": true
              },
              "unseenEpisodesCount": {
                "type": "number",
                "nullable": true
              },
              "userRating": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserRating"
                  }
                ],
                "nullable": true
              },
              "upcomingEpisode": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TvEpisode"
                  }
                ],
                "nullable": true
              },
              "lastAiredEpisode": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TvEpisode"
                  }
                ],
                "nullable": true
              },
              "nextAiring": {
                "type": "string",
                "nullable": true
              },
              "lastAiring": {
                "type": "string",
                "nullable": true
              },
              "onWatchlist": {
                "type": "boolean",
                "nullable": true
              },
              "lastSeenAt": {
                "type": "number",
                "nullable": true
              },
              "seen": {
                "type": "boolean",
                "nullable": true
              },
              "firstUnwatchedEpisode": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/TvEpisode"
                  }
                ],
                "nullable": true
              },
              "progress": {
                "type": "number",
                "nullable": true
              },
              "totalRuntime": {
                "type": "number",
                "nullable": true
              },
              "estimatedRating": {
                "type": "number",
                "nullable": true
              },
              "platformSeen": {
                "type": "boolean",
                "nullable": true
              },
              "metadataLanguage": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "List": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "privacy": {
            "$ref": "#/components/schemas/ListPrivacy"
          },
          "sortBy": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ListSortBy"
              }
            ],
            "nullable": true
          },
          "sortOrder": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ListSortOrder"
              }
            ],
            "nullable": true
          },
          "createdAt": {
            "type": "number"
          },
          "updatedAt": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          },
          "allowComments": {
            "type": "boolean",
            "nullable": true
          },
          "displayNumbers": {
            "type": "boolean",
            "nullable": true
          },
          "isWatchlist": {
            "type": "boolean"
          },
          "traktId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "createdAt",
          "id",
          "isWatchlist",
          "name",
          "privacy",
          "updatedAt",
          "userId"
        ]
      },
      "ListPrivacy": {
        "enum": [
          "friends",
          "private",
          "public"
        ],
        "type": "string"
      },
      "ListSortBy": {
        "enum": [
          "my-rating",
          "next-airing",
          "platform-recommended",
          "recently-added",
          "recently-aired",
          "recently-watched",
          "recommended",
          "release-date",
          "runtime",
          "title"
        ],
        "type": "string"
      },
      "ListSortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "type": "string"
      },
...
      "ListDetailsResponse": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "number"
              },
              "isWatchlist": {
                "type": "boolean"
              },
              "name": {
                "type": "string"
              },
              "createdAt": {
                "type": "number"
              },
              "updatedAt": {
                "type": "number"
              },
              "traktId": {
                "type": "number",
                "nullable": true
              },
              "sortOrder": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ListSortOrder"
                  }
                ],
                "nullable": true
              },
              "description": {
                "type": "string",
                "nullable": true
              },
              "privacy": {
                "$ref": "#/components/schemas/ListPrivacy"
              },
              "sortBy": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ListSortBy"
                  }
                ],
                "nullable": true
              },
              "allowComments": {
                "type": "boolean",
                "nullable": true
              },
              "displayNumbers": {
                "type": "boolean",
                "nullable": true
              }
            },
            "required": [
              "createdAt",
              "id",
              "isWatchlist",
              "name",
              "privacy",
              "updatedAt"
            ]
          },
          {
            "type": "object",
            "properties": {
              "totalRuntime": {
                "type": "number"
              },
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "username"
                ]
              }
            },
            "required": [
              "totalRuntime",
              "user"
            ]
          }
        ]
      },
      "ListItemsResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number"
            },
            "listedAt": {
              "type": "string"
            },
            "estimatedRating": {
              "type": "number",
              "nullable": true
            },
            "type": {
              "enum": [
                "audiobook",
                "book",
                "episode",
                "movie",
                "season",
                "tv",
                "video_game"
              ],
              "type": "string"
            },
            "mediaItem": {
              "$ref": "#/components/schemas/MediaItemItemsResponse"
            },
            "season": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TvSeason"
                }
              ],
              "nullable": true
            },
            "episode": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TvEpisode"
                }
              ],
              "nullable": true
            }
          },
          "required": [
            "id",
            "listedAt",
            "mediaItem",
            "type"
          ]
        }
      },
...
    }
  }
}